make s3 client work with simulation
