+++ fdbclient/include/fdbclient/S3Client.actor.h
@@ -26,9 +26,32 @@
 #elif !defined(FDBCLIENT_S3CLIENT_ACTOR_H)
 #define FDBCLIENT_S3CLIENT_ACTOR_H
 
-#include "fdbclient/Knobs.h"
-#include "fdbclient/SystemData.h"
+#include "fdbclient/S3BlobStore.h"
 #include "flow/actorcompiler.h" // This must be the last #include.
+
+// Copy files and directories to and from s3.
+// Has a main function so can exercise the actor from the command line. Uses
+// the S3BlobStoreEndpoint to interact with s3. The s3url is of the form
+// expected by S3BlobStoreEndpoint:
+//   blobstore://<access_key>:<secret_key>@<endpoint>/resource?bucket=<bucket>, etc.
+// See the section 'Backup URls' in the backup documentation,
+// https://apple.github.io/foundationdb/backups.html, for more information.
+// TODO: Handle prefix as a parameter on the URL so can strip the first part
+// of the resource from the blobstore URL.
+
+// TLS and blob credentials for backups and setup for these credentials.
+// Copied from fdbbackup/BackupTLSConfig.* and renamed S3CpTLSConfig.
+struct S3CpTLSConfig {
+	std::string tlsCertPath, tlsKeyPath, tlsCAPath, tlsPassword, tlsVerifyPeers;
+	std::vector<std::string> blobCredentials;
+
+	// Returns if TLS setup is successful
+	bool setupTLS();
+
+	// Sets up blob crentials. Add the file specified by FDB_BLOB_CREDENTIALS as well.
+	// Note this must be called after g_network is set up.
+	void setupBlobCredentials();
+};
 
 ACTOR Future<Void> copyDownDirectory(std::string s3url, std::string dirpath);
